examples/kubeflow-mlops/kubeflow-manifests/upstream/tests/e2e/requirements.txt
kubernetes==21.7.0
kfp==1.8.4
kubeflow-katib==0.12.0
